Refine creation helpers utilities and dtype one-value trait
